refactor [#227] SpotifyAPIHandler 의존성 분리 (DI)
